08_final_project/app/requirements.txt
bcrypt==4.2.1
catboost==1.2.7
fastapi==0.115.8
pandas==2.2.3
passlib==1.7.4
pika==1.3.2
pydantic==2.3.0
pytest==8.3.4
requests==2.32.3
SQLAlchemy==2.0.31
sqlmodel==0.0.22
streamlit==1.42.2
uvicorn==0.34.0
python-dotenv==1.0.0
pika==1.3.2
psycopg2-binary==2.9.7
joblib==1.5.1
scikit-learn==1.6.1
